time slider fixed but not styled
