+++ client/components/time-slider.js
@@ -1,10 +1,12 @@
-import React from 'react'
+import React, {useState, useEffect} from 'react'
 import {connect} from 'react-redux'
 import {makeStyles} from '@material-ui/core/styles'
 import IconButton from '@material-ui/core/IconButton'
 import {PlayCircleOutline} from '@material-ui/icons'
 import {Slider} from '@material-ui/lab'
 import Typography from '@material-ui/core/Typography'
+import marks from '../timeMarks'
+import {changeTimeThunk} from '../store'
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -24,71 +26,35 @@ const useStyles = makeStyles(theme => ({
   }
 }))
 
-const marks = [
-  {
-    value: 0,
-    label: '12 AM - 2 AM'
-  },
-  {
-    value: 9.09,
-    label: '2AM - 4AM'
-  },
-  {
-    value: 18.18,
-    label: '4AM - 6AM'
-  },
-  {
-    value: 27.27,
-    label: '6AM - 8AM'
-  },
-  {
-    value: 36.36,
-    label: '8AM - 10AM'
-  },
-  {
-    value: 45.45,
-    label: '10AM - 12PM'
-  },
-  {
-    value: 54.54,
-    label: '12PM - 2PM'
-  },
-  {
-    value: 63.63,
-    label: '2PM - 4PM'
-  },
-  {
-    value: 72.72,
-    label: '4PM - 6PM'
-  },
-  {
-    value: 81.81,
-    label: '6PM - 8PM'
-  },
-  {
-    value: 90.9,
-    label: '8PM - 10PM'
-  },
-  {
-    value: 99.99,
-    label: '10PM - 12AM'
-  }
-]
-
 function valueLabelFormat(value) {
   return marks.findIndex(mark => mark.value === value) + 1
 }
 
+const getHourPct = currentHour => {
+  let hours = marks.map(el => el.value)
+  let currentHourPct = currentHour / 24 * 100
+  return hours.filter(
+    el => currentHourPct >= el && currentHourPct < el + 9.09
+  )[0]
+}
+
+const getHour = pct => {
+  let hours = [0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22]
+  let hour = Math.floor(24 * (pct / 100))
+  return hours.filter(el => hour >= el && hour < el + 2)[0]
+}
+
 function DiscreteSlider(props) {
+  const [hourPct, setHourPct] = useState(getHourPct(props.mapView.currentHour))
+  useEffect(
+    () => {
+      props.changeTime(getHour(hourPct))
+    },
+    [hourPct]
+  )
+  let pct = hourPct
+
   const classes = useStyles()
-  const setHour = currentHour => {
-    let hours = marks.map(el => el.value)
-    let currentHourPct = currentHour / 24 * 100
-    return hours.filter(
-      el => currentHourPct > el && currentHourPct < el + 9.09
-    )[0]
-  }
-  let hourPct = setHour(props.mapView.currentHour)
   return (
     <div className={classes.root}>
       <Typography id="discrete-slider-restrict" gutterBottom>
@@ -99,13 +65,12 @@ function DiscreteSlider(props) {
           disabled={hourPct > 99}
           size="medium"
           onClick={() => {
-            if (hourPct < 99) {
+            if (pct < 99) {
               let int = setInterval(() => {
-                if (hourPct > 91) clearInterval(int)
-                hourPct = hourPct + 9.09
-                props.changeTime(Math.round(24 * (hourPct / 100)))
-                console.log(hourPct)
-              }, 2500)
+                if (pct > 90) clearInterval(int)
+                pct = pct + 9.09
+                setHourPct(pct)
+              }, 2000)
             }
           }}
         >
@@ -117,21 +82,12 @@ function DiscreteSlider(props) {
           className={classes.slider}
           valueLabelFormat={valueLabelFormat}
           aria-labelledby="discrete-slider-restrict"
+          min={0}
+          max={99.99}
           step={null}
           marks={marks}
+          onChange={(e, v) => setHourPct(v)}
           value={hourPct}
-          onChangeCommitted={(e, v) => {
-            let hourOnSlider = Math.round(v / 100 * 24)
-            let currentHour = Math.round(24 * (hourPct / 100))
-            if ([0, 2, 4].includes(hourOnSlider)) {
-              hourOnSlider += 1
-            }
-            if ([20, 22, 24].includes(hourOnSlider)) {
-              hourOnSlider -= 1
-            }
-            if (hourOnSlider !== currentHour)
-              return props.changeTime(hourOnSlider)
-          }}
         />
       </div>
     </div>
@@ -143,4 +99,10 @@ const mapStateToProps = state => {
   return {mapView}
 }
 
-export default connect(mapStateToProps, null)(DiscreteSlider)
+const mapDispatchToProps = dispatch => {
+  return {
+    changeTime: newTime => dispatch(changeTimeThunk(newTime))
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(DiscreteSlider)
