+++ client/timeMarks.js
@@ -1,0 +1,52 @@
+const marks = [
+  {
+    value: 0,
+    label: '12 AM - 2 AM'
+  },
+  {
+    value: 9.09,
+    label: '2AM - 4AM'
+  },
+  {
+    value: 18.18,
+    label: '4AM - 6AM'
+  },
+  {
+    value: 27.27,
+    label: '6AM - 8AM'
+  },
+  {
+    value: 36.36,
+    label: '8AM - 10AM'
+  },
+  {
+    value: 45.45,
+    label: '10AM - 12PM'
+  },
+  {
+    value: 54.54,
+    label: '12PM - 2PM'
+  },
+  {
+    value: 63.63,
+    label: '2PM - 4PM'
+  },
+  {
+    value: 72.72,
+    label: '4PM - 6PM'
+  },
+  {
+    value: 81.81,
+    label: '6PM - 8PM'
+  },
+  {
+    value: 90.9,
+    label: '8PM - 10PM'
+  },
+  {
+    value: 99.99,
+    label: '10PM - 12AM'
+  }
+]
+
+export default marks
